V1.2.0 - extract DeleteConfirmation
